Move formatWalletAddress to lib/utils
