+++ src/app/wallet/components/utils.tsx
@@ -2,10 +2,6 @@ import { TableTransaction } from '@/app/wallet/components/transaction-table';
 import { env } from '@/lib/env';
 import { Transaction } from '@/types/transactions';
 
-export const formatWalletAddress = (account: string): string => {
-  return `${account.slice(0, 5)}...${account.slice(-5)}`;
-};
-
 // TODO: Merge this with the one in @/lib/currency.ts?
 export const formatCurrency = (currency: string): string => {
   return Number(currency).toLocaleString();
